move transformWorkspaceEdit to transform.ts
